fix: 🐛 Negative balance issue fix for asset holders
When tokens for an Asset were issued/redeemed by an external agent,
those tokens were getting mapped with the Asset owner instead of
the external agent which was messing up the balance for asset holders.
This fix maps those tokens with the external agents and corrects the
existing data as well.